move HashContext into internal
