+++ provider/internal/testdata/dockerfile-location-irrelevant/app/foo.sh
