+++ provider/internal/testdata/dockerfile-location-irrelevant/step1.Dockerfile
