+++ provider/internal/testdata/dockerfile-location-irrelevant/step2.Dockerfile
