+++ provider/internal/testdata/filemode-matters/step1/Dockerfile
