+++ provider/internal/testdata/filemode-matters/step2-chmod-x/foo.sh
