+++ provider/internal/testdata/ignores-wildcard/basedir-modified-ignored-file/Dockerfile
