+++ provider/internal/testdata/ignores-wildcard/basedir-modified-ignored-file/projects/foo/src/app.js
