+++ provider/internal/testdata/ignores-wildcard/basedir-modified-included-file/projects/foo/src/app.js
