+++ provider/internal/testdata/ignores-wildcard/basedir/.dockerignore
