+++ provider/internal/testdata/ignores-wildcard/basedir/Dockerfile
