+++ provider/internal/testdata/ignores-wildcard/basedir/projects/foo/src/app.js
