+++ provider/internal/testdata/ignores/basedir-with-ignored-files/.dockerignore
