+++ provider/internal/testdata/ignores/basedir-with-ignored-files/Dockerfile
