+++ provider/internal/testdata/ignores/basedir-with-ignored-files/bar/app.js
