+++ provider/internal/testdata/ignores/basedir-with-ignored-files/foo.txt
