+++ provider/internal/testdata/ignores/basedir/Dockerfile
