+++ provider/internal/testdata/ignores/basedir/bar/app.js
