+++ provider/internal/testdata/ignores/basedir/foo.txt
