+++ provider/internal/testdata/renaming-matters/step1/Dockerfile
