+++ provider/internal/testdata/renaming-matters/step1/foo.a.txt
