+++ provider/internal/testdata/renaming-matters/step2/Dockerfile
