+++ provider/internal/testdata/renaming-matters/step2/foo.b.txt
