+++ provider/internal/testdata/symlinks/Dockerfile
