+++ provider/internal/testdata/symlinks/sub/dir/linkedFromTop.txt
