+++ provider/internal/testdata/symlinks/sub/dir/linkedToTop.txt
