+++ provider/internal/testdata/unignores/basedir-with-unignored-files/.dockerignore
