+++ provider/internal/testdata/unignores/basedir-with-unignored-files/Dockerfile
